+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # Next Release
 
 - [#380](https://github.com/IAMconsortium/pyam/pull/380) Add compatibility with latest matplotlib and py3.8
+- [#393](https://github.com/IAMconsortium/pyam/pull/393) Import ABC from collections.abc for Python 3.10 compatibility.
 
 # Release v0.6.0